0.10.0 Intermediate version before 1.0.0 (#5)
