+++ lib/puppet/util/http_client.rb
@@ -6,8 +6,8 @@ module Puppet
   module Util
     # Http client
     class HttpClient
-      FATAL_RESPONSE_CODES = Set.new(%w(400 401 402 403 405 406 407 409 410 411 412 413 414
-                                        415 416 417 422 423 424 426 428 431 501 502 505 507 511)).freeze
+      FATAL_RESPONSE_CODES = Set.new(%w[400 401 402 403 405 406 407 409 410 411 412 413 414
+                                        415 416 417 422 423 424 426 428 431 501 502 505 507 511]).freeze
 
       VERB_MAP = {
         get: Net::HTTP::Get,
