--- spec/defines/services/systemd_spec.rb
@@ -1,85 +1,0 @@
-# frozen_string_literal: true
-
-require 'spec_helper'
-
-describe 'graphdb::service::systemd', type: :define do
-  let :facts do
-    {
-      kernel: 'Linux',
-      operatingsystem: 'Ubuntu',
-      operatingsystemmajrelease: '16',
-      graphdb_java_home: '/opt/jdk8',
-      ipaddress: '129.10.1.1'
-    }
-  end
-
-  let(:title) { 'test' }
-
-  let :pre_condition do
-    "class { 'graphdb': version => '9.10.1', edition => 'ee' }"
-  end
-
-  context 'with ensure set to present' do
-    let(:params) { { ensure: 'present', service_ensure: 'running', service_enable: true } }
-
-    it do
-      is_expected.to contain_file('/lib/systemd/system/graphdb-test.service')
-        .with(ensure: 'present',
-              owner: 'root',
-              group: 'root',
-              before: "Service[graphdb-#{title}]",
-              notify: ['Exec[systemd_reload_test]', 'Service[graphdb-test]'])
-    end
-
-    it do
-      is_expected.to contain_file('/lib/systemd/system/graphdb-test.service')
-        .with(content: /Description="GraphDB - test"/)
-    end
-    it do
-      is_expected.to contain_file('/lib/systemd/system/graphdb-test.service')
-        .with(content: /Group=graphdb/)
-    end
-    it do
-      is_expected.to contain_file('/lib/systemd/system/graphdb-test.service')
-        .with(content: /User=graphdb/)
-    end
-    it do
-      is_expected.to contain_file('/lib/systemd/system/graphdb-test.service')
-        .with(content: /JAVA_HOME=\/opt\/jdk8/)
-    end
-    it do
-      is_expected.to contain_file('/lib/systemd/system/graphdb-test.service')
-        .with(content: /ExecStart=\/opt\/graphdb\/dist\/bin\/graphdb -Dgraphdb.home=\/opt\/graphdb\/instances\/test/)
-    end
-
-    it { is_expected.to contain_exec('systemd_reload_test').with(command: '/bin/systemctl daemon-reload', refreshonly: true) }
-
-    it do
-      is_expected.to contain_service('graphdb-test').with(
-        ensure: 'running',
-        enable: true,
-        name: 'graphdb-test.service',
-        provider: 'systemd',
-        hasstatus: true
-      )
-    end
-  end
-  context 'with ensure set to absent' do
-    let(:params) { { ensure: 'absent', service_ensure: 'disabled', service_enable: false } }
-
-    it do
-      is_expected.to contain_file('/lib/systemd/system/graphdb-test.service').with(ensure: 'absent', subscribe: 'Service[graphdb-test]')
-    end
-
-    it do
-      is_expected.to contain_service('graphdb-test').with(
-        ensure: 'disabled',
-        enable: false,
-        name: 'graphdb-test.service',
-        provider: 'systemd',
-        hasstatus: true
-      )
-    end
-    it { is_expected.to contain_exec('systemd_reload_test').with(command: '/bin/systemctl daemon-reload', refreshonly: true) }
-  end
-end
