+++ types/ensure.pp
@@ -1,0 +1,2 @@
+# @summary Validate the value of the ensure parameter for a install
+type Graphdb::Ensure = Enum['present', 'absent']
