+++ types/status.pp
@@ -1,0 +1,2 @@
+# @summary Validate the value of the status parameter for a install
+type Graphdb::Status = Enum['enabled', 'disabled', 'running', 'unmanaged']
